requirements.txt
requests>=2.25.0
Pillow>=8.0.0
imageio>=2.9.0
beautifulsoup4>=4.9.0
argparse>=1.4.0
